updated OCR to read vertical text
