+++ README.md
@@ -10,7 +10,7 @@ This repository holds some experiments I did in summer 2013 during a sudden inte
 
 Usage
 -----
-Provide an input file or path using the arguments `-i` and `-p`, respectively. The user may provide a corresponding output file or path using the `-o` argument. The output is saved in the default file/directory if output argument is defined.  
+Provide an input file or path using the arguments `-i` and `-p`, respectively. When an input path is provided, all files in the directory are processed. The user may provide a corresponding output file or path using the `-o` argument. The output is saved in the default file/directory if no output argument is defined.
 ```
     python MangaOCR_dev.py -i <infile> -o <outfile>
     python MangaOCR_dev.py -p <inpath> -o <outpath>
@@ -19,7 +19,7 @@ Provide an input file or path using the arguments `-i` and `-p`, respectively. T
 Other useful arguments:
 ```
     --html : Display output in an html file.
-    --default-directory : Store output in predefined folders.
+    --default-directory : Store output in predefined folders. This is recommended for processing images in a directory since the input may be overwritten.
     --additional_filtering : Attempt to filter false text positives by histogram processing.
 ```
 
@@ -34,17 +34,17 @@ I haven't bothered to form this into a python library. It's just a series of scr
 Text Location Example
 ---------------------
 Here's an example run of a page from Weekly Young Magazine #31 2013. The input image is as follows (jpg).
-![Input image](https://github.com/johnoneil/MangaTextDetection/blob/master/test/194.jpg?raw=true)
+![Input image](doc/194.jpg)
 
-An initial estimate of text locations can be found by the 'LocateText.py' script:
+An initial estimate of text locations can be found by the `MangaOCR.py` script:
 
 ```
- ../LocateText.py '週刊ヤングマガジン31号194.jpg' -o 194_text_locations.png
+    python MangaOCR.py -i 194.jpg
 ```
 
 With the results as follows (estimated text marked with red boxes):
 
-![locate text output](https://github.com/johnoneil/MangaTextDetection/blob/master/test/194_text_locations_thumb.png?raw=true)
+![Locate text output](doc/194_text_areas_no_filter.jpg)
 
 Note that in the output above you see several of the implementation deficiencies. For example, there are several small false positives scattered around, and some major false positives on the girl's sleeve and eyes in panels 2 and 3.
 Also note that many large areas of text were not detected (false negatives). Despite how pleased I was with the results (and I was more pleased than you could possibly believe) significant improvements are needed.
@@ -54,46 +54,22 @@ Text Segmentation Example
 To more easily separate text from background you can also segment the image, with text areas and non text being separated into different (RGB) color channels. This easily allows you to remove estimated text from image entirely or vice-versa.
 Use the command:
 ```
-./segmentation.py '週刊ヤングマガジン31号194.jpg' -o 194_segmentation.png
+    python segmentation.py 194.jpg -o 194_segmentation.png
 ```
 The results follow:
 
-![Input image](https://github.com/johnoneil/MangaTextDetection/blob/master/test/194_segmentation_thumb.png?raw=true)
+![Segmentation image](doc/194_segmentation.png)
 
 OCR and Html Generation
 -----------------------
-I did take the time to run simple OCR on some of the located text regions, with mixed results. I used the python tesseract package (pytesser) but found the results were not generally good for vertical text, among other issues.
-The script ocr.py should run ocr on detected text regions, and output the results to the command line.
-```
-../ocr.py '週刊ヤングマガジン31号194.jpg'
-Test blob assigned to no row on pass 2
-Test blob assigned to no row on pass 3
-0,0 1294x2020 71% :ぅん'・ 結局
-玉子かけご飯が
-一 番ぉぃしぃと
+I did take the time to run simple OCR on some of the located text regions. This uses the python wrapper for Google's OCR engine Tesseract (pytesserract). Unlike the original version of the `ocr.py` script, text is assumed to be vertically aligned.
 
-从
-胤
-赫
-囃
-包
-け
-H」
-の
-も
-側
-鵬
+I also embedded those results in an HTML output, allowing "readers" to hover on Japanese Text, revealing the OCR output, which can be edited/copied/pasted. This is by using the optional argument `--html`. Some examples can be seen below:
 
-はフィクショ穴ぁり、 登場する人物
+![OCR output success](doc/194_html_ocr_success.png)
+![OCR output success](doc/194_html_ocr_failure.png)
 
-※この物語
-
-```
-You can see some fragmented positives, but in all the results for this page are abysmal.
-
-I also embedded those results in an HTML output, allowing "readers" to hover on Japanese Text, revealing the OCR output, which can be edited/copied/pasted. This is via the script MangaDetectText. A (more successful) example of this can be seen below:
-
-![locate text output](https://github.com/johnoneil/MangaTextDetection/blob/master/test/example.png?raw=true)
+Even after using `jpn_vert` trained data, results are still unreliable. Most of the bounding boxes with larger width return successful results. Error mostly happens when the bounding box only has one line of vertically-aligned text.
 
 Dependencies
 -----------------------
@@ -102,12 +78,6 @@ You should be able to install most of the dependencies via pip, or you could use
 ### Python 3+
 
 https://www.python.org/
-
-Install as per OS instructions.
-
-### Pip
-
-http://pip.readthedocs.org/en/latest/index.html
 
 Install as per OS instructions.
 
@@ -156,8 +126,9 @@ Install as per OS instructions, this should also include the python bindings.
 https://code.google.com/p/tesseract-ocr/
 
 Install as per OS instructions, then use pip to install the python bindings.
-Don't forget to include your target language's trained data sets.
+Don't forget to include your target language's trained data sets and set the `tesseract_cmd` PATH. (See `ocr.py` script)
 
 ```
-pip install pytesseract
+Edit line 7 of ocr.py script:
+pytesseract.pytesseract.tesseract_cmd = <path>
 ```
