+++ defaults.py
@@ -25,8 +25,8 @@ FURIGANA_HORIZONTAL_SIGMA_MULTIPLIER=0.0
 FURIGANA_BINARY_THRESHOLD=240
 MINIMUM_TEXT_SIZE_THRESHOLD=7
 MAXIMUM_VERTICAL_SPACE_VARIANCE=5.0
-OCR_LANGUAGE='jpn'
-TEXT_DIRECTION=5
+OCR_LANGUAGE='jpn_vert'
+TEXT_DIRECTION=1
 IMAGE_OUTPUT_DIRECTORY='images/'
 TEXT_OUTPUT_DIRECTORY='text/'
 PATH_TO_HTML_TEMPLATE='./annotorious/template.html'
