+++ doc/ocr_results/194.html
@@ -1,0 +1,60 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8"/>
+        <style>
+            html, body {
+                padding:20px;
+                margin:0px;
+                font-family:'Lato', sans-serif;
+            }
+            #content {
+                width:100%;
+            }
+            h1 {
+                font-size:21px;
+                font-weight:normal;
+                margin:0;
+                padding:0;
+            }
+            img {
+                max-width:100%;
+            }
+        </style>
+        <link href="annotorious/annotorious.min.css" rel="stylesheet">
+        <script type="text/javascript" src="annotorious/annotorious.min.js"></script>
+    </head>
+    <body>
+        <div id="content">        
+            <h1>OCR Results</h1>    
+            <img id='image' src='194.jpg'>
+        </div>
+    </div>
+        <script>
+            var anno = Annotorious.init({
+                image: 'image',
+                locale: 'auto',
+            });
+            function addAnnotation_(x,y,w,h,text) {
+                var annotation = {
+                    "type": "Annotation",
+                    "body": {
+                        "type": "TextualBody",
+                        "value": text
+                    },
+                    "target": {
+                        "selector": [{
+                            "type": "FragmentSelector",
+                            "conformsTo": "http://www.w3.org/TR/media-frags/",
+                            "value": `xywh=pixel:${x},${y},${w},${h}`
+                        }]
+                    }
+                }
+                anno.addAnnotation(annotation)
+            }
+            for (tb of [{x:1216,y:122,w:78,h:552,text:" "},{x:240,y:172,w:94,h:212,text:" "},{x:917,y:178,w:10,h:136,text:" "},{x:475,y:495,w:8,h:8,text:" "},{x:581,y:581,w:27,h:19,text:" "},{x:431,y:993,w:16,h:21,text:"   "},{x:1202,y:1024,w:92,h:890,text:"   "},{x:16,y:1059,w:953,h:32,text:"   1S 潜 那 吾 \ ゝ NYm\d い で ・ 過 帆 呈 ひ > 藻 ・ 回 非 然 記 周 SeWShCu ル つつ 温 京 に ご 中 いい  "},{x:47,y:1110,w:94,h:209,text:"た ぶん お 母さん に 。 言っ て も 分 か ん な いよ ::;  "},{x:360,y:1111,w:57,h:106,text:"え ~ っ と  "},{x:490,y:1149,w:34,h:143,text:"する の 2?  "},{x:534,y:1149,w:81,h:181,text:"埋 練 て どん な 練習  "},{x:1045,y:1151,w:141,h:207,text:"うん ! 結 局  玉子 か け ご 飯 が 一 番 お いし いと 思う ! 言い  "},{x:256,y:1208,w:55,h:16,text:"   "},{x:179,y:1342,w:30,h:39,text:" "},{x:49,y:1433,w:91,h:156,text:"で ちょ  "},{x:1004,y:1488,w:13,h:33,text:" "},{x:1056,y:1505,w:59,h:150,text:"(ルー が //  "},{x:331,y:1656,w:57,h:133,text:": 邊 曲 と 六 ビー  ・。・・  "},{x:1099,y:1660,w:9,h:11,text:"//  "},{x:529,y:1661,w:95,h:181,text:"て いう か あん た  何 か 叫 ん で な か つた ルル  "},{x:184,y:1669,w:66,h:145,text:"いて きま ーー すり:  "},{x:62,y:1674,w:47,h:107,text:"フ  ロロ ロ  "}]) {
+                addAnnotation_(tb.x,tb.y,tb.w,tb.h,tb.text)
+            }
+        </script>
+    </body>
+</html>
